Added ViewRecipe()
Removed the ClearData() method as it is no longer needed in the case of my program, I have added a ViewRecipe method that allows the user to enter the recipe name, and have the corresponding recipe be displayed.

removed the parsing of variables in DisplayRecipe, to prevent the need for variable passing in the ViewRecipe() method.